DTFS2-4362 - split out notify template formatters
